requirements.txt
Django==2.2.11
django-filter==2.2.0
django-import-export==1.2.0
django-jchart==0.4.2
django-model-utils==3.2.0
django-quiz-app==0.5.1
gunicorn==19.9.0
Pillow==6.2.2
